APIMF-2997: added corrections for JS bundle

+++ shared/src/main/scala/amf/client/remod/amfcore/config/ParsingOptions.scala
@@ -5,7 +5,7 @@ package amf.client.remod.amfcore.config
   */
 case class ParsingOptions(amfJsonLdSerialization: Boolean = true,
                           baseUnitUrl: Option[String] = None,
-                          maxYamlReferences: Option[Long] = None) {
+                          maxYamlReferences: Option[Int] = None) {
 
   /** Parse specific AMF JSON-LD serialization */
   def withoutAmfJsonLdSerialization: ParsingOptions = copy(amfJsonLdSerialization = false)
@@ -20,10 +20,10 @@ case class ParsingOptions(amfJsonLdSerialization: Boolean = true,
   def withoutBaseUnitUrl(): ParsingOptions = copy(baseUnitUrl = None)
 
   /** Defines an upper bound of yaml alias that will be resolved when parsing a DataNode */
-  def setMaxYamlReferences(value: Long): ParsingOptions = copy(maxYamlReferences = Some(value))
+  def setMaxYamlReferences(value: Int): ParsingOptions = copy(maxYamlReferences = Some(value))
 
-  def isAmfJsonLdSerialization: Boolean  = amfJsonLdSerialization
-  def definedBaseUrl: Option[String]     = baseUnitUrl
-  def getMaxYamlReferences: Option[Long] = maxYamlReferences
+  def isAmfJsonLdSerialization: Boolean = amfJsonLdSerialization
+  def definedBaseUrl: Option[String]    = baseUnitUrl
+  def getMaxYamlReferences: Option[Int] = maxYamlReferences
 
 }
